+++ lib/ServerlessRuntimeBase.js
@@ -1,8 +1,8 @@
 'use strict';
 
-const SError     = require('../ServerlessError'),
+const SError     = require('./ServerlessError'),
   BbPromise    = require('bluebird'),
-  SUtils       = require('../utils/index'),
+  SUtils       = require('./utils/index'),
   fs           = require('fs'),
   path         = require('path');
 
@@ -10,7 +10,7 @@ const SError     = require('../ServerlessError'),
  * This is the base class that all Serverless Runtimes should extend.
  */
 
-class ServerlessBaseRuntime {
+class ServerlessRuntimeBase {
   constructor(S, name) {
     this.S = S;
     this.name = name;
@@ -52,4 +52,4 @@ class ServerlessBaseRuntime {
   };
 }
 
-module.exports = ServerlessBaseRuntime;
+module.exports = ServerlessRuntimeBase;
