+++ lib/ServerlessRuntimeNode.js
@@ -1,13 +1,13 @@
 'use strict';
 
-const SError   = require('../ServerlessError'),
-  BaseRuntime  = require('./ServerlessBaseRuntime'),
-  SCli         = require('../utils/cli'),
-  SUtils       = require('../utils/index'),
+const SError   = require('./ServerlessError'),
+  RuntimeBase  = require('./ServerlessRuntimeBase'),
+  SCli         = require('./utils/cli'),
+  SUtils       = require('./utils/index'),
   _            = require('lodash'),
   BbPromise    = require('bluebird');
 
-class ServerlessNodeRuntime extends BaseRuntime {
+class ServerlessRuntimeNode extends RuntimeBase {
   constructor(S) {
     super( S, 'nodejs' );
   }
@@ -37,4 +37,4 @@ class ServerlessNodeRuntime extends BaseRuntime {
   }
 }
 
-module.exports = ServerlessNodeRuntime;
+module.exports = ServerlessRuntimeNode;
