+++ lib/ServerlessRuntimePython27.js
@@ -1,8 +1,8 @@
 'use strict';
 
-const SError     = require('../ServerlessError'),
-  SCli         = require('../utils/cli'),
-  BaseRuntime  = require('./ServerlessBaseRuntime'),
+const SError     = require('./ServerlessError'),
+  SCli         = require('./utils/cli'),
+  RuntimeBase  = require('./ServerlessRuntimeBase'),
   BbPromise    = require('bluebird'),
   path         = require('path'),
   _            = require('lodash'),
@@ -19,7 +19,7 @@ function pipPrefixInstall(requirements, dir) {
   process.chdir(process.cwd());
 };
 
-class ServerlessPython27Runtime extends BaseRuntime {
+class ServerlessRuntimePython27 extends RuntimeBase {
   constructor(S) {
     super( S, 'python2.7' );
   }
@@ -55,4 +55,4 @@ class ServerlessPython27Runtime extends BaseRuntime {
   }
 }
 
-module.exports = ServerlessPython27Runtime;
+module.exports = ServerlessRuntimePython27;
