+++ lib/utils/index.js
@@ -25,8 +25,8 @@ BbPromise.promisifyAll(fs);
  */
 
 module.exports.supportedRuntimes = {
-  "nodejs": require('../runtimes/ServerlessNodeRuntime'),
-  "python2.7": require('../runtimes/ServerlessPython27Runtime')
+  "nodejs": require('../ServerlessRuntimeNode'),
+  "python2.7": require('../ServerlessRuntimePython27')
 };
 
 /**
